fix check plugin
